# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/tag-categories": {
      "get": {
        "deprecated": false,
        "operationId": "AdminController_getTagCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all tag categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagCategoryWithTagsDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tag categories",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "deprecated": false,
        "operationId": "AdminController_createTagCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tag category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTagCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new tag category",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/tag-categories/{id}": {
      "put": {
        "deprecated": false,
        "operationId": "AdminController_updateTagCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag category updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTagCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a tag category",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "deprecated": false,
        "operationId": "AdminController_deleteTagCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag category deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a tag category",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/tags": {
      "get": {
        "deprecated": false,
        "operationId": "AdminController_getTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tags",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "deprecated": false,
        "operationId": "AdminController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tag created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTagDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new tag",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/tags/{id}": {
      "put": {
        "deprecated": false,
        "operationId": "AdminController_updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTagDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a tag",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "deprecated": false,
        "operationId": "AdminController_deleteTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a tag",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games": {
      "get": {
        "deprecated": false,
        "operationId": "AdminController_getGames",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Title search, across supported translations.",
            "schema": {
              "example": "catan",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of games with translations and tags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminGamesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get games with translations and tags (paginated and searchable)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games/{id}/translations": {
      "put": {
        "deprecated": false,
        "operationId": "AdminController_updateGameTranslations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameTranslationsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game translations updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update game translations",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games/{id}/tags": {
      "put": {
        "deprecated": false,
        "operationId": "AdminController_updateGameTags",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameTagsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game tags updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update game tags",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/proposals": {
      "get": {
        "deprecated": false,
        "operationId": "AdminController_getGameProposals",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "duplicate"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of game proposals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminGameProposalsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get game proposals (paginated and filterable by status)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/proposals/{id}": {
      "get": {
        "deprecated": false,
        "operationId": "AdminController_getGameProposal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game proposal details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameProposalCompleteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific game proposal",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "deprecated": false,
        "operationId": "AdminController_deleteGameProposal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game proposal deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a game proposal",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/proposals/{id}/approve": {
      "post": {
        "deprecated": false,
        "operationId": "AdminController_approveGameProposal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveGameProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game proposal approved and game created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "createdGameId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a game proposal and create the game",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/proposals/{id}/reject": {
      "post": {
        "deprecated": false,
        "operationId": "AdminController_rejectGameProposal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectGameProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game proposal rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a game proposal",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/proposals/{id}/duplicate": {
      "post": {
        "deprecated": false,
        "operationId": "AdminController_markGameProposalAsDuplicate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "reviewNotes",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Already present in the catalog.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game proposal marked as duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a game proposal as duplicate",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Board Vault API",
    "description": "Versioned contract snapshot for the Board Vault social group workspace API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "The success status of the operation."
          }
        },
        "required": [
          "success"
        ]
      },
...
      "TagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Tag Name"
          },
          "categoryId": {
            "type": "number",
            "example": 1
          },
          "gameCount": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "id",
          "name",
          "categoryId",
          "gameCount"
        ]
      },
      "TagCategoryWithTagsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Action"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "gameCount": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "id",
          "name",
          "tags",
          "gameCount"
        ]
      },
      "CreateTagCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Action"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tag Name"
          },
          "categoryId": {
            "type": "number",
            "example": 1
          },
          "gameCount": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "name",
          "categoryId",
          "gameCount"
        ]
      },
...
      "AdminGamesResponseDto": {
        "type": "object",
        "properties": {
          "games": {
            "description": "Array of games with their translations and tags",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameWithTagsAndTranslationsDto"
            }
          },
          "pagination": {
            "description": "Pagination information",
            "allOf": [
              {
                "$ref": "#/components/schemas/BrowseGamesPaginationDto"
              }
            ]
          }
        },
        "required": [
          "games",
          "pagination"
        ]
      },
      "UpdateGameTranslationsBody": {
        "type": "object",
        "properties": {
          "en": {
            "type": "string",
            "description": "English translation of the game title",
            "example": "Catan"
          },
          "es": {
            "type": "string",
            "description": "Spanish translation of the game title",
            "example": "Catan"
          }
        }
      },
      "UpdateGameTagsBody": {
        "type": "object",
        "properties": {
          "tagIds": {
            "description": "Array of tag IDs to assign to the game",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "tagIds"
        ]
      },
      "GameProposalCompleteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "submittedBy": {
            "type": "number",
            "example": 1,
            "description": "Account ID that submitted the proposal"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "duplicate"
            ],
            "description": "Current status of the proposal"
          },
          "title": {
            "type": "string",
            "example": "Catan",
            "description": "Game title"
          },
          "imageUrl": {
            "type": "object",
            "example": "https://www.example.com/image.jpg",
            "nullable": true,
            "description": "Game image URL"
          },
          "gameAvgDuration": {
            "type": "object",
            "example": 120,
            "nullable": true,
            "description": "Average duration of the game in minutes"
          },
          "minPlayers": {
            "type": "object",
            "example": 3,
            "nullable": true,
            "description": "Minimum number of players required"
          },
          "maxPlayers": {
            "type": "object",
            "example": 4,
            "nullable": true,
            "description": "Maximum number of players allowed"
          },
          "proposedTags": {
            "type": "object",
            "example": "[\"strategy\", \"family\"]",
            "nullable": true,
            "description": "Proposed tags as JSON string"
          },
          "notes": {
            "type": "object",
            "example": "A great family strategy game",
            "nullable": true,
            "description": "User notes about the game"
          },
          "reviewedBy": {
            "type": "object",
            "example": 2,
            "nullable": true,
            "description": "Admin account ID that reviewed the proposal"
          },
          "reviewedAt": {
            "type": "object",
            "example": "2024-01-15T10:30:00Z",
            "nullable": true,
            "description": "When the proposal was reviewed"
          },
          "reviewNotes": {
            "type": "object",
            "example": "Approved - great game for our collection",
            "nullable": true,
            "description": "Admin review notes"
          },
          "createdGameId": {
            "type": "object",
            "example": 123,
            "nullable": true,
            "description": "ID of the created game if approved"
          },
          "submittedAt": {
            "type": "string",
            "example": "2024-01-10T15:45:00Z",
            "description": "When the proposal was submitted"
          },
          "submitterId": {
            "type": "number",
            "example": 1,
            "description": "ID of the user who submitted the proposal"
          },
          "reviewerId": {
            "type": "number",
            "example": 2,
            "nullable": true,
            "description": "ID of the admin who reviewed the proposal"
          }
        },
        "required": [
          "id",
          "submittedBy",
          "status",
          "title",
          "imageUrl",
          "gameAvgDuration",
          "minPlayers",
          "maxPlayers",
          "proposedTags",
          "notes",
          "reviewedBy",
          "reviewedAt",
          "reviewNotes",
          "createdGameId",
          "submittedAt",
          "submitterId",
          "reviewerId"
        ]
      },
      "AdminGameProposalsResponseDto": {
        "type": "object",
        "properties": {
          "proposals": {
            "description": "Array of game proposals with submitter and reviewer information",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameProposalCompleteDto"
            }
          },
          "pagination": {
            "description": "Pagination information",
            "allOf": [
              {
                "$ref": "#/components/schemas/BrowseGamesPaginationDto"
              }
            ]
          }
        },
        "required": [
          "proposals",
          "pagination"
        ]
      },
      "ApproveGameProposalBody": {
        "type": "object",
        "properties": {
          "reviewNotes": {
            "type": "string",
            "example": "Great game for our collection!",
            "description": "Admin review notes for approval"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://www.example.com/game-image.jpg",
            "description": "Image URL for the game (if not provided in proposal)"
          },
          "gameAvgDuration": {
            "type": "number",
            "example": 120,
            "description": "Average duration in minutes (if not provided in proposal)"
          },
          "minPlayers": {
            "type": "number",
            "example": 3,
            "description": "Minimum players (if not provided in proposal)"
          },
          "maxPlayers": {
            "type": "number",
            "example": 4,
            "description": "Maximum players (if not provided in proposal)"
          },
          "translations": {
            "type": "object",
            "example": {
              "en": "Catan",
              "es": "Catan"
            },
            "description": "Game title translations"
          },
          "tagIds": {
            "example": [
              1,
              2,
              3
            ],
            "description": "Tag IDs to assign to the game",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}